+++ UPGRADE.md
@@ -639,10 +639,10 @@ find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Normal
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Normalizer\\Standard\\FamilyNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Standard\\FamilyNormalizer/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Normalizer\\Standard\\FamilyVariantNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Standard\\FamilyVariantNormalizer/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Normalizer\\Storage\\AttributeOptionNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Storage\\AttributeOptionNormalizer/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\AttributeGroupNormalize/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versionning\\AttributeGroupNormalizer/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\AttributeNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versionning\\AttributeNormalizer/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\AttributeOptionNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versionning\\AttributeOptionNormalizer/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\FamilyNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versionning\\FamilyNormalizer/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\AttributeGroupNormalize/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versioning\\AttributeGroupNormalizer/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\AttributeNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versioning\\AttributeNormalizer/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\AttributeOptionNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versioning\\AttributeOptionNormalizer/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\FamilyNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versioning\\FamilyNormalizer/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\ApiBundle\\Controller\\AttributeController/Akeneo\\Pim\\Structure\\Bundle\\Controller\\ExternalApi\\AttributeController/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\ApiBundle\\Controller\\AttributeGroupController/Akeneo\\Pim\\Structure\\Bundle\\Controller\\ExternalApi\\AttributeGroupController/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\ApiBundle\\Controller\\AttributeOptionController/Akeneo\\Pim\\Structure\\Bundle\\Controller\\ExternalApi\\AttributeOptionController/g'
@@ -727,7 +727,7 @@ find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\Arra
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Akeneo\\Tool\\Component\\Api\\Normalizer\\AssociationTypeNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\ExternalApi\\AssociationTypeNormalizer/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Normalizer\\AssociationTypeNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\InternalApi\\AssociationTypeNormalizer/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Normalizer\\Standard\\AssociationTypeNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Standard\\AssociationTypeNormalizer/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\AssociationTypeNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versionning\\AssociationTypeNormalizer/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\VersioningBundle\\Normalizer\\Flat\\AssociationTypeNormalizer/Akeneo\\Pim\\Structure\\Component\\Normalizer\\Versioning\\AssociationTypeNormalizer/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Repository\\AssociationTypeRepositoryInterface/Akeneo\\Pim\\Structure\\Component\\Repository\\AssociationTypeRepositoryInterface/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\EnrichBundle\\Controller\\Rest\\GroupTypeController/Akeneo\\Pim\\Structure\\Bundle\\Controller\\InternalApi\\GroupTypeController/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Connector\\ArrayConverter\\FlatToStandard\\GroupType/Akeneo\\Pim\\Structure\\Component\\ArrayConverter\\FlatToStandard\\GroupType/g'
