+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/serializers_versioning.yml
