+++ src/Akeneo/Pim/Structure/Component/Normalizer/Versioning/AttributeNormalizer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Pim\Structure\Component\Normalizer\Versionning;
+namespace Akeneo\Pim\Structure\Component\Normalizer\Versioning;
 
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
