+++ src/Akeneo/Pim/Structure/Component/spec/Normalizer/Versioning/AttributeNormalizerSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Akeneo\Pim\Structure\Component\Normalizer\Versionning;
+namespace spec\Akeneo\Pim\Structure\Component\Normalizer\Versioning;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
@@ -8,7 +8,7 @@ use Akeneo\Pim\Structure\Component\Model\AttributeOption;
 use Akeneo\Pim\Structure\Component\Model\AttributeOptionValue;
 use Akeneo\Pim\Enrichment\Component\Product\Normalizer\Versioning\TranslationNormalizer;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
-use Akeneo\Pim\Structure\Component\Normalizer\Versionning\AttributeNormalizer;
+use Akeneo\Pim\Structure\Component\Normalizer\Versioning\AttributeNormalizer;
 use Prophecy\Argument;
 
 class AttributeNormalizerSpec extends ObjectBehavior
